requirements.txt
onnxruntime-gpu
opencv-python
numpy
tqdm
realesrgan
basicsr
